e2e refactoring (#2298)

+++ tests/e2e/clearstorage.helpers.js
@@ -5,21 +5,23 @@ export async function clearStorage() {
   //await global.context.clearCookies();
   //await global.client.clearStorageState();
 
-  // Evaluate JavaScript to clear local storage.
-  await global.client.evaluate(() => {
-    window.localStorage.clear();
-  });
-  await global.client.evaluate(() => {
-    window.sessionStorage.clear();
-  });
-  // Execute JavaScript to clear cookies.
-  await global.client.evaluate(() => {
-    document.cookie.split(';').forEach(function (c) {
-      document.cookie = c
-        .replace(/^ +/, '')
-        .replace(/=.*/, '=;expires=' + new Date().toUTCString() + ';path=/');
+  if (global.client) {
+    // Evaluate JavaScript to clear local storage.
+    await global.client.evaluate(() => {
+      window.localStorage.clear();
     });
-  });
+    await global.client.evaluate(() => {
+      window.sessionStorage.clear();
+    });
+    // Execute JavaScript to clear cookies.
+    await global.client.evaluate(() => {
+      document.cookie.split(';').forEach(function (c) {
+        document.cookie = c
+          .replace(/^ +/, '')
+          .replace(/=.*/, '=;expires=' + new Date().toUTCString() + ';path=/');
+      });
+    });
+  }
   /*if (global.session) {
     //await global.client.evaluate(() => window.localStorage.clear());
     //await global.client.evaluate(() => window.sessionStorage.clear());
